fix(perf): Fully fallback transaction summary based on cardinality (#50173)
### Summary
Previously we were using the 'setting' for metrics fallback decision
which could be propagated from the landing page. There are too many
cases where this doesn't happen and we still have flows from entirely
'null' transactions that currently don't work with the clusterer
whatsoever, this means **showing the transaction summary based on
metrics data will not work.** This does a full fallback similarly to how
the landing page does. We can remove it in the future when we're closer
to 100% ingestion and landing no longer needs the fallback either.

This should fix the cases of "no transactions" appearing in either the
graph or the transaction list _due to transaction.source:unknown_

![Screenshot 2023-06-01 at 12 40 43
PM](https://github.com/getsentry/sentry/assets/6111995/fcd7fa6c-09ec-401b-8909-fba313f39795)