ref(group-preview-tooltip): Remove broken event previews (#50332)
The group preview tooltip previously accepted either a group ID or an
event ID + project slug. The idea was that you could also preview
specific events without a corresponding group. However, since the
addition of performance issues we require the issue category to know
what kind of preview to display. Ever since that addition, we can't know
how to display an event preview and the event previews have been broken.

This just formalizes the need for an issue category and removes all the
code paths that use event ID + project slug.