feat(options) Prevent drift when updating through different channels (#49744)
The options automator is going to be a new tool that allows to make
changes to runtime options in a way that supports multiple regions and
single tenants, guarantees accountability and allows for reviews in
specific cases. It does that by keeping the source of truth in a github
repo that is automatically synchronized with production by a GoCD
pipeline.

As part of the Options Automator project we will have two sources of
truth to update options for a period of time.
This can cause drift in this scenario:
- an option is set in the DB via the CLI or via the UI admin
- the option is set to a different value via the Automator by storing
the value in the options git repo.
- We do not want to let the option updated from outside the DB to
overwrite the value on the DB until the Automator is the sole source of
truth.

For this we need to know when an drift between the two sources is
happening.
The previous PR introduced a `last_updated_by` column in the options
table. This allows us, when we try to update an option, to know whether
the option in the DB has drifetd, which means it has been assigned a
different value through a different channel.

Having the `last_updated_column` in the DB allows us to prevent
overwrite selectively. For example we would allow the Automator to
overwrite the Admin UI soon.

This PR adds a method to the options manager to evaluate the update
policies. It takes into account these scenarios:
- whether the option can be updated at all (we have readonly ones)
- whether the option cannot be updated by anybody in the current state
(options that are overridden by settings)
- whether the option has the flag that would allow a specific
UpdateChannel to update it (Automator cannto update credentials)
- whether the option has drifted

It also adds a check in the `register` method to prevent an option to be
assigned conflicting flags. Like readonly and automator_updatable.