ref(dropdownMenu): Update `trigger` prop type (#50424)
Update `DropdownMenu`'s `trigger` prop type from
```tsx
  trigger?: (
    props: Omit<React.HTMLAttributes<Element>, 'children'> & {
      onClick?: (e: MouseEvent) => void;
    }
  ) => React.ReactNode;
```
to
```tsx
  trigger?: (
    props: Omit<React.HTMLAttributes<HTMLElement>, 'children'>,
    isOpen: boolean
  ) => React.ReactNode;
```
The main change here is in the `isOpen` prop: it's been moved outside
the `props` parameter and is now it's own standalone parameter in the
render function. Doing this helps prevent typing conflicts in the future
(and paves the way for an upcoming `@react-aria/button` package
upgrade). `isOpen` is obviously not a native HTML prop — it's only being
used by `DropdownButton`. Isolating `isOpen` from the `props` parameter
means we can pass `props` to other components (e.g. a plain `Button`)
that don't accept `isOpen` as a prop.