fix(trends): add trendsParameter to query to show correct parameter graph (#50305)
<!-- Describe your PR here. -->
Issue that was brought up in @udameli bug bash where switching the
trends parameter didn't switch the trends graph.
### Examples

![image](https://github.com/getsentry/sentry/assets/72356613/68a4a0d9-1f79-4b8e-aa3e-a408baba9e00)

![image](https://github.com/getsentry/sentry/assets/72356613/38d29b62-e0b9-44f1-bca1-585c924cce2d)

### The Fix
The graph using metrics data gets this information via the
trendsParameter which was not being updated in the query. The
trendsParameter is now visible in the query and is being updated as the
user changes the parameter. Now the the graph gets updated with the
correct data when the trend parameter is changed.