fix: Metrics indexer drops any inflight messages on join
Any messages in the processing pipeline prior to the produce step
should be dropped.

Previously this timeout applied to the ParallelTransform step, which
would wait 10 seconds for join to complete. However, since the subsequent
step (Unbatch) drops all messages on join anyway, it is pointless to parallel
transform messages that will later be dropped.