ref(ingest): Create an arroyo-based ingest consumer (#49957)
Adds a simplistic arroyo-based variant of the ingest consumers. This
consumer is fully sequential without batching its messages, and it
delegates to the same functions as the former consumer did.

To enable the new variant, pass --v2-consumer as argument to the CLI.
The new consumer no longer supports listening to multiple topics, so
each of the three ingest consumers must be started separately with this
flag enabled. Do not enable parallelism, as this may lead to data
inconsistencies.

---------

Co-authored-by: Jan Michael Auer <mail@jauer.org>
Co-authored-by: Markus Unterwaditzer <markus-tarpit+git@unterwaditzer.net>