feat(starfish) Fetch span summary information from indexed spans (#50492)
The span operation and description on the span summary page can be
fetched from production data now! The way it works is that given a span
group, it fetches exactly 1 row from the indexed spans table with that
span group. It is _technically possible_ for more than one span
description to correspond to a single span group, but I don't think it's
critical to cover that case for now.

I had to introduce a new `IndexedSpan` type, since right now the field
names are different from span metrics. Hopefully we'll have time to make
them match. Otherwise, this is mostly all the same code, just with
slightly different fields.