feat(generic-metrics): Use case-aware cardinality rate limiter (#49121)
https://getsentry.atlassian.net/browse/SNS-2228

Now that each message in the metrics MessageProcessor has a use case
(`use_case_id`) associated with it, we want to be able to set up rate
limits on a per use case-basis (i.e. have different configurations for
different use cases such as `spans` and `escalating_issues` and
`transactions`).

This PR attempts to introduce the changes we need in order to apply rate
limit configs on a per use case-basis. The next step would be to add
these configs/options and use them in this code.

---------

Co-authored-by: John Yang <john.yang@sentry.io>