feat(options automator). Update the batch script to account for drift. (#50280)
https://github.com/getsentry/sentry/pull/49744 introduces the logic to
detect drift between options set via the traditional channels and the
new options automator.
Traditional channels are the admin UI and the CLI. The automator uses a
configMap as source of truth. So drift can happen.

This updates the `configoptions` script to take the changes of the PR
above into account.
Specifically the changes are:
- removed all the commands that were not used or that were already
present in the old config CLI
- removed the concept of TRACKED_OPTIONS. We have a flag that tells when
an option should be managed by the automator.
- changes the schema of the configmap file to be just a large dictionary
of option keys/values. No need to separate it in sections.
- Two command remain: patch and sync. Patch takes the content of the
file and applies it to the DB. It ignores options that are not in the
file. Sync takes the file as a source of truth. All the options not
present in the file are unset if they are set in the DB.
- Validates whether all options are not read only before executing any
update. If the file is malformed, options are not registered, invalid
options are present the script bails immediately.
- Skips the updates of drifted options. If the option is drifted but
otherwise valid, the script does not fail. The individual drifted
options are skipped.