feat(plugins): Add discord webhook plugin
Add discord webhook plugin that uses the vanilla webhook[1].

While the discord slack-compatible webhook[2] and Sentry legacy plugin
work together in principle, it is not possible to utilise the `custom_message` field
to trigger mentions in Discord, since Discord show the entire message in
a single embed including the `custom_message` field content,
whereby any scheme of mention is disabled (@here, @everyone, <@userid>,
@<#channelid>).

With the vanilla webhook, we can preserve the embed in a similar format
and add the `custom_message` field content as part of a normal message
(as opposed to being part of the embed), whereby all scheme of mention works.

This commit is almost a one-to-one translation of the legacy Slack
plugin, colours are translated from hex to decimal with https://www.spycolor.com/,
field names have changed to adapt to using the vanilla webhook. Further
functionality such as thread creation, customisable embed colour, user
interaction with buttons, etc. are out of scope in this commit.

[1] - https://discord.com/developers/docs/resources/webhook#execute-webhook
[2] - https://discord.com/developers/docs/resources/webhook#execute-slackcompatible-webhook