ref(replays): remove user fields from subquery eligibility (#50615)
Helps with https://github.com/getsentry/sentry/issues/48957 -- this will
re-allow user fields to work correctly. We solved OOM problems with
https://github.com/getsentry/snuba/pull/4232, so it's unlikely this
would result in bad queries.



Note that there is a better fix which is doing a distinct query instead,
which we would need to add another clickhouse query option for, but this
would work jsut fine.