ref(ingest): Move to arroyo consumer by default (#50481)
This changes the `ingest-consumer` command to use the arroyo consumer by
default. Legacy options still remain in the CLI for the time being so
that we can migrate all usages over.

Addionally, all tests relying on the legacy ingest consumer now use the
new arroyo consumer. This means that the last remaining uses of
`BatchingKafkaConsumer` and related utilities have been removed from the
code base.