feat(api): Return 404 from internal Relay endpoint (#50716)
Sentry routes a number of endpoints to Relay. This does not include an
internal endpoint at `/api/relay/*`, which is meant to inspect Relay
itself. Such an endpoint is useful for operating Relay, but does not
belong to the public API of Sentry.

In Sentry's web API, this path falls back to a catch-all endpoint, which
redirects to a login view. This is undesirable for the Relay endpoint,
especially if someone tries to access `/api/relay/healthcheck/live/` and
receives a redirect response.

This PR ensures we respond with 404 on all paths that are assigned to
Relay.