ref(sentry_metrics): Use sentry.utils.arroyo to configure subprocesses.
We now have a subclass of RunTaskWithMultiprocessing that, in theory,
automatically copies global metric tags into each subprocess, and
initializes sentry-the-django-app for you.

Additionally, do not pass already-global tags to the MetricsWrapper.
Global tags did not work correctly in the past (see introduction of
sentry.metrics.middleware) and this was necessary to get any tags to
show up in the dashboards.