fix(hc): sentry_app.api_application_id can be None, outboxes should work (#50750)
SentryApp's relation to ApiApplication can be set null on cascaded
deletion. I suppose also it's possible for the api application
underlying SentryApp to change in some cases (which seems weird!)

That said, we order sentry app's outboxes based on the api application
to ensure application level consistency. In the case that we do have a
null here, though, it is correct and fine to order then by `0` which
will never be a real identifier.