feat(options) Make the options automator render the value of drifted options  (#50692)
The expected workflow when we encounter drifted options (the option on
the DB has been changed outside of the option automator configmap) is
that:
- the automator will skip the update and yell.
- we would update the configmap to the value contained in the DB
- we would try again to apply the change. Now the script would notice
that we are not changing the value but the the previous DB update was
not done by the automator.
- At this point the automator leaves the value in place (there is no
more drift in the value between configmap and DB) but switches the last
update field to automator in order to be able to make further changes.

The previous version of the script made this process hard because it,
well, did not print out the value present in the DB when drift was
detected. That would have made it really hard to fix the configmap file.
This PR adds the option (enabled by default) to display the DB value
when we encounter drift.