fix(rulesnooze): Change text in email (#50826)
this pr changes the text in the email for muting alerts to match the
changes made in https://github.com/getsentry/sentry/pull/49573/ . Now,
clicking mute will essentially "click" whatever the default mute button
is for that alert (which depends on the notifications send for the
alert). For example, clicking "Mute this alert" for an alert that just
sends the default notification will only mute for the user.