fix(hybridcloud) Fix a few pydantic warning sources
We were passing none into `user_ids` filter conditions which was
causing serialization errors.

The change in the release serializer is because when we're serializing
commits (and their releases) we don't pass the current request user
through the various layers of logic and response serializers invent an
Anonymous user in this condition. The AnonymousUser cannot be converted
into an RpcUser.

Refs SENTRY-106Q