ref(replays): remove unused method on replay File model (#50849)
as mypy points out, there is no `name` property on this model. This
method was copied over originally when I copy/pasted the code from
EventAttachment to make this model. we do not use this method anywhere,
and should be fine to remove. Fixes the typing issue in this file.