Add Service monitoring
This changes the existing queue-based monitoring to a more generic `service` monitoring.

Services are organized in a hierarchy of base services, and consumers depending on those.

The main monitor / monitoring loop will check the health of different base services,
and aggregates that health for the consumers. The consumer health is then persisted and queried.

The settings around monitoring and checking are also streamlined, and a new setting for high watermarks replaces the existing queue-size based options.