feat(ts): Add LinkButton alias
This is the ground work for extracting a LinkButton from Button.

In cases where a Button has a `to` or `href` attribute, these are
actually anchor (`a`) elements. We want to differentiate this at the
component level to simplify the props, and also fix some improper usages
(such as `<Button priority="link" href="..." />`, which should just be a
`Link`).

After this change Buttons using `to` and `href` are deprecated. Instead
you should use `LinkButton` which in the future will be properly typed
as an anchor