fix(perf-issues): Require space in N+1 ext span description (#50994)
Continuing the experiment to remove the query parameterization check
from the N+1 DB Query detector, I noticed that as a side effect we
started detecting issues for Mongo N+1s, which isn't appropriate. The
simplest thing we could do: check for a space in the query. (Everything
we care about will definitely have at least one space in it.)

This change is for experimentation (only affects metrics collected and
doesn't affect any actual issue creation yet).

Will add a test if this change graduates to the real N+1 DB detector.