docs(api): Update openapi-examples-validator (#51019)
Update Openapi-examples-validator to version 5.0.0

The previous version 4.1.1 had a bug that was only fixed in 4.7.0 that incorrectly validated schema with additional Properties when the `--no-additional-properties` flag was set https://github.com/codekie/openapi-examples-validator/issues/178. Shoutout to @mdtro for discovering this in an earlier [PR](https://github.com/getsentry/sentry/pull/43216#issuecomment-1396261406)

Upgrading also forces the `nullable` property to explicitly be used with a type, so we include one fix for that.