feat(hybrid-cloud): Use ProxyClient for Jira requests (#51005)
This PR adds the ProxyClient for requests to Jira. This will be
necessary since the hostname of the server is used to generate the JWT
that's set to Jira. Requests with a JWT generated on the region silos
will fail, so now those requests are proxied through the control.

Two other notable changes
- Small refactors done to the base classes for better logging when
debugging and adding compatibility with query params
- JWT is now sent as an `Authorization` header rather than a query param
so we don't expose it in server logs (even if only valid for 5m at a
time)

**todo**
- [x] Add tests

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>