fix(release-filter): Apply dashboard saved filters to widget viewer
When the dashboard is saved, the unsaved filters get saved to state and
removed from the URL. We weren't taking into account the saved state so
filters weren't being applied. Instead, we should pass along filters to
the modal and the dashboard should be responsible for fetching the
filters from the URL or state.