feat(js): Add `endpoint` tag to all request errors (#50910)
This is a follow-up to https://github.com/getsentry/sentry/pull/49558, which added an `endpoint` tag to all error events with a request error as their cause. In this PR, the same tag is added to events whose primary error is a request error, by using a helper function in `beforeSend`.