feat(generic-metrics): Kafka-based metrics interface (#47897)
Initial pass on the metrics interface with the Kafka-based backend. It
doesn't return Futures for now, but this should be a follow-up item.

At some point, there will also be a Relay-based metrics interface, which
could be adapted to return Futures as well (I think). That interface
would likely have to write some logic for parsing the HTTP response from
Relay and instantiating/handling Futures accordingly.