fix: Use multiprocessing strategy in ingest consumer tests
The multiprocessing strategy in sentry.utils.arroyo automatically
disables multiprocessing in tests. So it should now be safe to increase
processes, because it will be ignored anyway.

I believe this is the right boundary to test at, because in principle
anything that makes multiprocessing behave differently from regular
RunTask is a bug in Arroyo.

In exchange, add some validation that the passed initializer and
function arguments are pickleable.