fix(txnames): Prevent scan of all redis keys (#50940)
Default environments do not define a dedicated redis cluster or redis
instance for the transaction name clusterer. Since we use `scan_iter` to
iterate through all keys in Redis, the clusterer becomes slower the
larger the cluster is overall, independent of how many keys we actually
hit.

To prevent scanning all keys, introduce a new "meta" set of project IDs.
Whenever a set of samples is newly created for a project, its ID is
added to this "meta" set in redis. When spawning clusterers, we iterate
only over this set instead.

This PR also unifies some of the helper functions of transaction name
and span description clustering, such that the change affects both.