fix(org-tokens): Ensure all users can update token names
This was actually specced out different than I initially implemented this, which I noticed just now - so fixing this to how it _should_ be. Note that as of now you can only update the token name, which is very low-risk, which is why we allow this for everyone (otherwise, users could not adjust the name of tokens they created). In the future, we _may_ adjust this so that users can only edit tokens they created themselves, but for now this is fine.