fix(events-facets): Use the pagination helper (#51177)
The initial PR to expose page size and cursor allowed us to paginate,
but it's missing data about whether or not there are more results. The
pagination helper puts that information in the response header so we can
inspect it in the frontend and know if there's more data.

This PR is mostly a whitespace change because it wraps the original data
fetching logic in a function that gets called with varying `limit` and
`offset` values, derived from the `cursor` provided.