feat(hybrid-cloud): Allow changing historic monolith region name (#51192)
This PR adds a function to the split database script which overwrites
your historic monolith mappings with whatever you've set in your
settings. It can be rerun with specifying the historic name as well, in
case it wasn't `--monolith--`.

It also updates the silo docs with this new pattern of running sentry.