ref(crons): Remove sort button temporarily (#51297)
<img width="323" alt="image"
src="https://github.com/getsentry/sentry/assets/9372512/46cc85c5-6940-412c-8d92-7cd69e68ea66">

Button on the left was added in according to mocks and put behind a
flag, but we haven't implemented the sort re-ordering functionality and
the flag is now turned on. Let's hide this for now until we implement
the sort.