fix(hybridcloud) Add missing routing for django.contrib (#51259)
There are a few adjacent and related fixes in these changes:

1. We need to be able to run migrations for a few django.contrib models.
These models have foreign keys with other models so I've assigned them
all to control silo.
2. I've changed the connection alias from `region` to `default` for the
region silo database to reduce the number of `transaction.atomic()`
blocks that will need to be changed. Instead of having to change them
all we only need to change control silo operations.
3. I've fixed the `post_migrate` signal handlers from running on *each*
connection that is migrated. Instead the signal handlers only do work
when the connection for the relevant model is migrated.

With all of these fixes in place I've been able to get a handful of
model tests to pass with `SENTRY_USE_SILOS=1 pytest [path]`

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>