fix(crons): Filter out alert rule if no targets specified (#51317)
Currently if a user makes no selection here 
<img width="761" alt="image"
src="https://github.com/getsentry/sentry/assets/9372512/0ada0afc-97d1-465a-a655-ee4ded146202">
we still send an empty `{}` to the backend, which results in a validator
error. We should allow users to not have to specify environments or
alert rule, so filtering it out in the `transformData` function.