feat(events-facets): Account for project tag injection in pagination
When querying for multiple projects, the project tag gets injected. This
is a problem for pagination because the injection removes a result from
the first page and so the other pages have an offset that needs to be
reduced by one, or else that removed result is always omitted.

If we're requesting subsequent pages, subtract 1 from the offset to
account for this. Also appends project at the beginning of the results
so it isn't truncated by the paginator.