feat(starfish): consume spans-samples endpoint (#51296)
This PR makes a bunch of changes in order to consume the spans-samples
endpoint in the span sample list sidebar. It also fixes some bugs and
does a few refactor items to make it all work. It still needs some work,
but it's almost there!

![image](https://github.com/getsentry/sentry/assets/44422760/b960ef62-c5e3-408b-ba1a-12c9fbb1e248)

Here's a list of changes
1. Consume span-samples endpoint
2. Fix bug where date in span samples duration chart would show 1970.
3. Refactor to use the shared `SpanIndexedFieldTypes`

Future TODO
1. Fix `Load more samples` button - it will load the exact same samples!
I think it's the endpoint not randomizing it everytime, you might need
to wait a bit between requests to get a different result.
2. See if the design needs some changes.