feat(starfish): Allow sorting the span table (#51301)
- `useSpanList` now accepts a `sorts` prop instead of `orderby`. I'm not
100% sold on this, but it's kind of nice, that way a `Sort` object is
used everywhere, which makes it possible to have great typing. Try
giving `SpansTable` a sort on a field it doesn't support and watch it
warn you! Pretty good
- fix it so `SpansTable` doesn't maintain any weird sorting state,
instead using the URL query parameter `spansSort`. I think that's kind
of gross, but it's what we're doing with the cursors for now, so I did
the same thing with sorts
- improve the typing in `SpansTable`. Much less duplication, and now
it'll throw errors when you try to feed it data it doesn't have, and so
on

---------

Co-authored-by: Dominik Buszowiecki <44422760+DominikB2014@users.noreply.github.com>