fix(flamegraph): Filter system frames when building profile (#51253)
Lifts the frame filtering logic to when the profile is built. This means
that the frame merging logic is correctly applied when the same frame
appears adjacent to each other. This also ensures to adjust the config
space when in flamegraph mode to the weight of the flamegraph so hidden
frames are not taking up horizontal space.