feat(starfish): Split out span view table delta columns (#51393)
The "Change" columns are now separate, have their own headings, and are
sortable. I ended up doing this the "right" way, which is to hook up
response meta, and the field renderer pipeline. Unfortunately, our
TypeScript types for the responses are _wrong_, so there's still some
work to do, but this is a good first step! From now on we can do a lot
more of this kind of reusable rendering.