fix(migrations) Fix incorrectly folded keys in squash migration (#51413)
When I was compiling the squash migration I missed several foreign key
folds. This has blocked generating the split database from the sqaush
migration as foreign keys are still present in the migration operation
list.

These changes do a few things:

1. Fold operations that would result in a HybridCloudForeign key into
the create table operation.
2. Fold intermediary states and index/constraint operations that would
change a relation field into the CreateModel operation.

There is also a state modification to change the user bigint back into an int.
This is a workaround as the model definition for User is inconsistent with
the physical schema in production. I plan to correct this with an upcoming migration.