ref(replay): Refactor useReplayData and create a new wrapper useReplayReader (#51536)
The return type of `useReplayData` always bugged me a bit. It was
returning almost all the raw data (errors and replayRecord, but not
attachments) and also all the hydrated data (a ReplayReader instance).

It should be doing one thing only!

So this introduces a little wrapper, `useReplayReader` which takes all
the raw data from `useReplayData` and gets a Reader instance back out.

It's also a good place to parse the `replaySlug` into `[projectSlug,
replayId]` parts. We'll need to keep this parsing for a while, until all
links in the app are fixed to use pass replayId, and then all old data
has expired.

Tests are also be simplified because the old file has fewer
dependencies, and the replaySlug stuff can be tested in isolation.