fix(events-facets): Increase aggregate tags limit
The default limit by the query builder is 50. We want to have at most 9
top values per tag key, but if there are tag keys with a lot of values
then we can hit the query builder limit before we've actually returned
values for every key, resulting in behaviour where the pages are missing
results.

Set the limit to the number of aggregate tags we're querying for
multiplied by the tag limit to ensure we're limiting the correct amount.