ref(hc): Rename variable in jira webhook (#51562)
This is a follow-up to https://github.com/getsentry/sentry/pull/50208/, which changed `get_integration_from_jwt` from returning an `Integration` instance to returning a `RpcIntegration` instance. That PR also renamed the variable pointing to the result of a `get_integration_from_jwt` call in the Jira uninstalled webhook. This PR applies the same renaming to the Jira issue-updated webhook, so the differences in what they're doing (one converts the `RpcIntegration` into an `Integration,` while the other doesn't) are clearer.

It also updates the docstring of `bind_org_context_from_integration` to reflect that fact that it may now be called with the id of an `RpcIntegration` instead of the id of an `Integration`.