chore: Update User.id descriptor to match physical schema (#51521)
Our production database is *already* using bigint so this is a state
only change.

The reason for this change is to ensure that future usage of
`FlexibleForeignKey` infers the correct field type when foreign keys are
created to `User`

I discovered this inconsistency when updating the
squash migration. To work around it, I made the physical schema use
bigint but model state use `int` at the end of the initial migration.
This workaround allowed the initial schema to be generated with bigint
but not have inconsistent state.