fix(codecov): Wait for extra job before making codecov PR comment (#51566)
This is a follow-up to https://github.com/getsentry/sentry/pull/51558, which increased the number of shards into which the backend test suite is broken in GHA. The extra shard means that codecov should now wait for an extra coverage report to be uploaded before it adds its comment to PRs. This fixes the codecov config and adds a note to both the backend and frontend GHA configs to future folks to make this change.