fix(perf-issues): MN+1 detector threshold should only consider db (#49998)
The MN+1 detector has a duration threshold to help ensure that only
meaningful problems are surfaced. Previously, the detector used all the
repeating spans when determining if the duration was over the threshold.
A customer reported an unexpected MN+1 where less than 1% of the time
was actually taken up by DB queries, and therefore wouldn't be
meaningfully impacted by fixing the N+1.

Update the MN+1 detector to only consider the time taken up by database
queries when checking the duration threshold.