feat(sdk): Add org context when instantiating `Pipeline` class (#51643)
This adds org context data to the SDK scope when creating `Pipeline` instances.  (A pipeline is used when walking a user through a multi-view flow, with state persisting between views, such as when signing in using GH or Google or when installing an integration.)

(This was inspired by work adding org context to integration endpoints. That works fine, except with an integration's `installed` webhook, since the integration isn't yet attached to an org at that point. While this doesn't solve that problem directly (the best we can do there is tag with the integration id), it does ensure that other code that runs during integration installation will have the correct data.)