ref(auth): Update ApiApplication serializer (ER-1448) (#51570)
`allowed_origins` and `redirect_uris` are stored as character fields in
our DB. I _think_ we accept both space separated and newline separated
lists. This PR updates our outbound serializers to have consistent
returns for these fields.

Note that we're using this model for a new endpoint for instance level
OAuth clients (i.e. to eventually allow users to sign in to CodeCov with
Sentry), but it's [also used for creating API Applications
](https://sentry.io/settings/account/api/applications/). The code for
this was written 6-8 years ago, and there might be subtle side effects
of this change somewhere.

Although this passes all tests and shouldn't be breaking, I'd feel
better if someone familiar with that part of our codebase/product could
take a look, if anyone knows who that may be please let me know!