feat(starfish): add bounds when fetching for span samples (#51545)
In order to prevent weird samples way higher then the p95 causing the
duration yAxis to change significantly. We want to maintain the
"quadrants" of the original p95 graph.

In addition to this change, I also had to update the `useSpansQuery` to
be dynamically typed. I need more properties from `useQuery`,
specifically `isFetching` as `isLoading` only applies to inital data
loads and not on manual refetches. This is also a better sol'n imo as we
get access to all the `useQueryResult` properties.