feat(hybridcloud) Add a label to trigger hybrid cloud builds
As we work towards having split databases be the norm for CI, we need
a way to incrementally improve our code compatibility without disrupting
master.

To achieve this, I've put together a pull request (#51252) to update our
CI automation to read this label and change the CI configuration for as
long as the label is applied. My plan is to use this to get incremental
progress reports as we fix errors in the current state. Once we have
all the tests passing, I'll add this to the CI configuration until
it is no longer 'feature flagged'.