fix(muting-metric-alerts): Prevent re-fetch of data when muting (#51829)
to mute metric alerts, we have a query parameter that we want to hide
after muting the alert. right now, getting rid of the query parameter
re-fetches the data which can potentially not show the most up to date
snooze information. this pr fixes the issue by not refetching if the
change to the url is because of muting.