ref(hc): Restructure RPC response body with "value" and "meta" fields
This makes responses symmetric to the request body (which has "args" and
"meta"), and prevents the awkwardness of having an empty response body
to represent a logical null return value.

As with requests, the "meta" field is currently empty but available for
future use. It also gives us the opportunity to add other fields to the
root-level object, rather than being stuck with the raw return value at
the root level.