fix(js-sdk): Move tracePropagationTargets to top level config (#51900)
In https://github.com/getsentry/sentry/pull/51797 we bumped the JS SDK
to `7.57.0`. Changelog here:
https://github.com/getsentry/sentry-javascript/releases/tag/7.57.0

Notable in this release was adding the ability to set
`tracePropagationTargets` as a top level option to `Sentry.init`.
Previously `tracePropagationTargets` was a nested option inside the
`BrowserTracing` integration.

```js
// old way
Sentry.init({
  integrations: [
   Sentry.BrowserTracing({
      tracePropagationTargets: ['foo', 'bar'],
    }),
  ],
});

// new with `7.57.0`
Sentry.init({
  integrations: [new Sentry.BrowserTracing()],
  // controls what requests trace headers attached - required because cors 😭
  tracePropagationTargets: ['foo', 'bar']
});
```

This change is done in a backwards compat manner though, so that if you
defined `tracePropagationTargets` in `BrowserTracing`, it is considered
valid. What's notable is that we made the decision to say that defining
a top level `tracePropagationTargets` (in `Sentry.init`) would override
the `tracePropagationTargets` set in `BrowserTracing` options - this
minimized the bundle size of the backwards compat change.

Unfortunately, our sdk init code in the sentry frontend was
unintentionally setting `tracePropagationTargets` in the top level, via
the spread operator.

```ts
sentryConfig as {
  allowUrls: string[];
  dsn: string;
  release: string;
  tracePropagationTargets: string[];
  profilesSampleRate?: number;
};

Sentry.init({
    // This defines tracePropagationTargets on the init options object
    ...sentryConfig,
// ...
```

Previously this did nothing, but with the bump to `7.57.0`, we started
to use this top level `tracePropagationTargets` instead of the one in
`BrowserTracing`. This means we removed the `/^\//` target in
`['localhost', /^\//, ...extraTracePropagationTargets]`, which meant we
stopped attaching headers to our outgoing api requests.

This PR fixes this predicament by removing the usage of `BrowserTracing`
`tracePropagationTargets` in favour of using the top level option.