ref: Remove default processing store, use redis backend everywhere
Redis event processing store is currently unused, meaning we can change
it to not use a cache backend and move forward with Ted's original
vision of getting rid of the cache backend.

getsentry already uses a "transitional" backend that does not use the
cache layer, and now that the redis backend is identical to that one we
can move over to it.

Previously I broke sentry by deploying
https://github.com/getsentry/sentry/pull/51870#issuecomment-1613493378,
where I forgot that getsentry had its own processingstore backend. I
could fix the backend but I can also just delete it.